Buggy 3D reconstruction

+++ schlam/cpp/src/data/gt_data.cpp
@@ -2,9 +2,9 @@
 #include "gt_data.hpp"
 
 GTData::GTData(const utils::TTimestamp &aTimestamp,
-                 const Eigen::Vector3d &aPosition,
-                 const Eigen::Matrix3d &aRotation,
-                 const Eigen::Vector3d &aTranslationalVelocity,
+                 const Eigen::Vector3f &aPosition,
+                 const Eigen::Matrix3f &aRotation,
+                 const Eigen::Vector3f &aTranslationalVelocity,
                  const std::string &aCF)
     : mTimestamp(aTimestamp), mPosition(aPosition), mRotation(aRotation),
       mTranslationalVelocity(aTranslationalVelocity), mCoordinateFrame(aCF) {}
