+++ schlam/cpp/src/data/gt_data.hpp
@@ -8,14 +8,14 @@
 class GTData {
 public:
   GTData(const utils::TTimestamp &aTimestamp,
-          const Eigen::Vector3d &aPosition, 
-          const Eigen::Matrix3d &aRotation,
-          const Eigen::Vector3d &aTranslationalVelocity,
+          const Eigen::Vector3f &aPosition,
+          const Eigen::Matrix3f &aRotation,
+          const Eigen::Vector3f &aTranslationalVelocity,
           const std::string &aCF);
 
   const utils::TTimestamp mTimestamp;
-  const Eigen::Vector3d mPosition;
-  const Eigen::Matrix3d mRotation;
-  const Eigen::Vector3d mTranslationalVelocity;
+  const Eigen::Vector3f mPosition;
+  const Eigen::Matrix3f mRotation;
+  const Eigen::Vector3f mTranslationalVelocity;
   const std::string mCoordinateFrame;
 };
