+++ schlam/cpp/src/data/image_data.cpp
@@ -2,7 +2,7 @@
 #include "image_data.hpp"
 
 ImageData::ImageData(const utils::TTimestamp &aTimestamp,
-                    cv::Mat &aImage, const Eigen::Matrix3d &aIntrinsics,
+                    cv::Mat &aImage, const Eigen::Matrix3f &aIntrinsics,
                     const std::string &aCF)
     : mTimestamp(aTimestamp)
     , mImage(aImage)
