+++ schlam/cpp/src/data/image_data.hpp
@@ -10,10 +10,10 @@
 class ImageData {
 public:
   ImageData(const utils::TTimestamp &aTimestamp, cv::Mat &aImage,
-            const Eigen::Matrix3d &aIntrinsics, const std::string &aCF);
+            const Eigen::Matrix3f &aIntrinsics, const std::string &aCF);
 
   const utils::TTimestamp mTimestamp;
   cv::Mat mImage;
-  const Eigen::Matrix3d mIntrinsics;
+  const Eigen::Matrix3f mIntrinsics;
   const std::string mCoordinateFrame;
 };
