+++ schlam/cpp/src/data/imu_data.hpp
@@ -8,16 +8,16 @@
 class IMUData {
 public:
   IMUData(const utils::TTimestamp &aTimestamp,
-          const Eigen::Vector3d &aAcceleration,
-          const Eigen::Vector3d &aAngularVelocity, const std::string &aCF,
+          const Eigen::Vector3f &aAcceleration,
+          const Eigen::Vector3f &aAngularVelocity, const std::string &aCF,
           const double mGyroscopeNoiseDensity,
           const double mGyroscopeRandomWalk,
           const double mAccelerometerNoiseDensity,
           const double mAccelerometerRandomWalk);
 
   const utils::TTimestamp mTimestamp;
-  const Eigen::Vector3d mAcceleration;
-  const Eigen::Vector3d mRotationalVelocity;
+  const Eigen::Vector3f mAcceleration;
+  const Eigen::Vector3f mRotationalVelocity;
   const std::string mCoordinateFrame;
   const double mGyroscopeNoiseDensity;
   const double mGyroscopeRandomWalk;
