+++ schlam/cpp/src/tft/main.cpp
@@ -18,15 +18,15 @@ int main()
    
     auto render_loop = std::thread(std::bind(&Plotter::run, &plotter));
 
-    Eigen::Matrix3d rotZ180{{-1, 0, 0}, {0, -1, 0}, {0, 0, 1}}; 
-    Eigen::Matrix3d rotZ90{{0, -1, 0}, {1, 0, 0}, {0, 0, 1}};
+    Eigen::Matrix3f rotZ180{{-1, 0, 0}, {0, -1, 0}, {0, 0, 1}};
+    Eigen::Matrix3f rotZ90{{0, -1, 0}, {1, 0, 0}, {0, 0, 1}};
 
     std::vector<std::shared_ptr<tft::RigidTransform3D>> transforms;
-    transforms.push_back(std::make_shared<tft::RigidTransform3D>("frame1", "world", rotZ180, Eigen::Vector3d(-1, -1, -1)));
-    transforms.push_back(std::make_shared<tft::RigidTransform3D>("frame2", "world", rotZ90, Eigen::Vector3d(3, 0, 0)));
-    transforms.push_back(std::make_shared<tft::RigidTransform3D>("frame3", "world", rotZ90, Eigen::Vector3d(1, 1, 1)));
-    transforms.push_back(std::make_shared<tft::RigidTransform3D>("frame4", "frame3", rotZ90, Eigen::Vector3d(1, 1, 1)));
-    transforms.push_back(std::make_shared<tft::RigidTransform3D>("frame5", "frame4", rotZ90, Eigen::Vector3d(1, 1, 1)));
+    transforms.push_back(std::make_shared<tft::RigidTransform3D>("frame1", "world", rotZ180, Eigen::Vector3f(-1, -1, -1)));
+    transforms.push_back(std::make_shared<tft::RigidTransform3D>("frame2", "world", rotZ90, Eigen::Vector3f(3, 0, 0)));
+    transforms.push_back(std::make_shared<tft::RigidTransform3D>("frame3", "world", rotZ90, Eigen::Vector3f(1, 1, 1)));
+    transforms.push_back(std::make_shared<tft::RigidTransform3D>("frame4", "frame3", rotZ90, Eigen::Vector3f(1, 1, 1)));
+    transforms.push_back(std::make_shared<tft::RigidTransform3D>("frame5", "frame4", rotZ90, Eigen::Vector3f(1, 1, 1)));
 
 
     for (auto transform : transforms) {
@@ -34,9 +34,9 @@ int main()
         
         plotter.addTransform(transform);
 
-        auto x = tft::Transformable3D(Eigen::Vector3d(1, 0, 0), transform->mSource);
-        auto y = tft::Transformable3D(Eigen::Vector3d(0, 1, 0), transform->mSource);
-        auto z = tft::Transformable3D(Eigen::Vector3d(0, 0, 1), transform->mSource);
+        auto x = tft::Transformable3D(Eigen::Vector3f(1, 0, 0), transform->mSource);
+        auto y = tft::Transformable3D(Eigen::Vector3f(0, 1, 0), transform->mSource);
+        auto z = tft::Transformable3D(Eigen::Vector3f(0, 0, 1), transform->mSource);
 
         std::cout << "x: " << x << std::endl;
         std::cout << "y: " << y << std::endl;
